+++ changelog.md
@@ -2,3 +2,8 @@
 - Added lock and unlock commands to reroll.
 - Added permission message when a player doesn't have permission to use the command.
 - Refactored all 3 reroll commands into a single command.
+- Remove useless aliases from reroll command.
+- Fix lock command feedback to send the player name rather than the player entity.
+- Change @Mod to use metadata from mcmod.info.
+- Update mcmod.info with valid information.
+- Add mod logo file.